+++ src/main/kotlin/com/wutsi/platform/core/cache/spring/CacheConfigurationNone.kt
@@ -1,0 +1,30 @@
+package com.wutsi.platform.core.cache.spring
+
+import org.springframework.beans.factory.annotation.Value
+import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty
+import org.springframework.cache.CacheManager
+import org.springframework.cache.support.NoOpCache
+import org.springframework.cache.support.SimpleCacheManager
+import org.springframework.context.annotation.Bean
+import org.springframework.context.annotation.Configuration
+
+@Configuration
+@ConditionalOnProperty(
+    value = ["wutsi.platform.cache.type"],
+    havingValue = "none",
+    matchIfMissing = true
+)
+open class CacheConfigurationNone(
+    @Value("wutsi.platform.cache.name") private val name: String
+) {
+    @Bean
+    open fun cacheManager(): CacheManager {
+        val cacheManager = SimpleCacheManager()
+        cacheManager.setCaches(
+            listOf(
+                NoOpCache(name)
+            )
+        )
+        return cacheManager
+    }
+}