+++ CHANGELOG.md
@@ -1,6 +1,16 @@
 # Change Log
-## 0.0.1 - 15-Aug-2021
-Initial
 
-## 0.0.1 - 15-Aug-2021
-Initial
+## [0.0.4] - 2021-08-15
+### Added
+- Add `caching` with memcached and local implementation
+
+### Changed
+- Set no-stream as default implementation
+- Set no-cache a default implementation
+- Set local storage a default implementation
+
+## [0.0.1] - 2021-08-15
+Initial version
+### Added
+- Initial version with `error`, `logging`, `storage`, `stream` and `tracing`
+